refactor(ks-vm-new): execute instructions directly from bytes

+++ workspace/ks-vm-new/src/types.rs
@@ -2,13 +2,13 @@ use crate::utils::VMError;
 
 pub type CaptureSize = usize;
 pub type Offset = i32;
-pub type VariableId = u64;
+pub type VariableId = u32;
 pub type Pointer = usize;
 
-pub type Slot = u64;
+pub type Slot = u32;
 
-pub type CollectionId = u64;
-pub type StorageId = u64;
+pub type CollectionId = u32;
+pub type StorageId = u32;
 pub type Owners = u32;
 
 pub type NativeId = usize;
